requirements.txt
aniso8601==1.0.0
Babel==1.3
blinker==1.4
Flask==0.10.1
Flask-Babel==0.9
Flask-Cache==0.13.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
flask-marshmallow==0.6.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.4
Flask-SQLAlchemy==2.0
Flask-Testing==0.4.2
Flask-WTF==0.11
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
marshmallow==2.0.0b4
marshmallow-sqlalchemy==0.2.0
passlib==1.6.5
python-dateutil==2.4.2
pytz==2015.4
requests==2.10.0
six==1.9.0
speaklater==1.3
SQLAlchemy==1.0.8
Unidecode==0.4.18
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
